Example: evaluating simple arithmetic expressions
We keep an external stack for the evaluation.

The external code has to know the specific rule numbers that are being
reduced at all points; this would be better achieved by creating a
JSON-like representation of the parser, and using that externally.

BUT: this assumes the parse is unambiguous. Doesn't tomita already have
a "stack" that can represent multiple parses? Need to look into this
with more detail.